+++ apps/app/package.json
@@ -7,11 +7,11 @@
     "dev:all": "concurrently \"pnpm dev\" \"npx inngest-cli@latest dev\" \"pnpm db:studio\"",
     "build": "next build --turbopack",
     "start": "next start",
-    "db:generate": "drizzle-kit generate",
-    "db:migrate": "drizzle-kit migrate",
-    "db:studio": "drizzle-kit studio",
-    "db:pull": "drizzle-kit pull",
-    "db:push": "drizzle-kit push",
+    "db:generate": "drizzle-kit generate --config ../../packages/db/drizzle.config.ts",
+    "db:migrate": "drizzle-kit migrate --config ../../packages/db/drizzle.config.ts",
+    "db:studio": "drizzle-kit studio --config ../../packages/db/drizzle.config.ts",
+    "db:pull": "drizzle-kit pull --config ../../packages/db/drizzle.config.ts",
+    "db:push": "drizzle-kit push --config ../../packages/db/drizzle.config.ts",
     "db:sync": "pnpm db:push && pnpm db:generate && pnpm db:migrate && pnpm db:pull",
     "inngest": "pnpm dlx inngest-cli@latest dev",
     "lint": "eslint .",
@@ -86,6 +86,7 @@
     "zod": "^4.3.6"
   },
   "devDependencies": {
+    "@pilot/config": "workspace:*",
     "@tailwindcss/postcss": "^4.2.0",
     "@types/node": "^24.10.13",
     "@types/react": "19.2.2",
