+++ apps/app/postcss.config.mjs
@@ -1,5 +1,1 @@
-const config = {
-  plugins: ["@tailwindcss/postcss"],
-};
-
-export default config;
+export { default } from "@pilot/config/postcss";
