+++ packages/config/eslint.config.mjs
@@ -1,0 +1,55 @@
+import typescriptEslint from "@typescript-eslint/eslint-plugin";
+import typescriptParser from "@typescript-eslint/parser";
+
+const eslintConfig = [
+  {
+    ignores: [
+      "node_modules/**",
+      ".next/**",
+      "out/**",
+      "build/**",
+      "next-env.d.ts"
+    ]
+  },
+  {
+    files: ["**/*.{js,jsx}"],
+    languageOptions: {
+      ecmaVersion: "latest",
+      sourceType: "module",
+      parserOptions: {
+        ecmaFeatures: {
+          jsx: true
+        }
+      }
+    },
+    rules: {
+      "prefer-const": "error",
+      "no-var": "error",
+      "no-unused-vars": "warn"
+    }
+  },
+  {
+    files: ["**/*.{ts,tsx}"],
+    languageOptions: {
+      parser: typescriptParser,
+      parserOptions: {
+        ecmaVersion: "latest",
+        sourceType: "module",
+        ecmaFeatures: {
+          jsx: true
+        },
+        project: "./tsconfig.json"
+      }
+    },
+    plugins: {
+      "@typescript-eslint": typescriptEslint
+    },
+    rules: {
+      "@typescript-eslint/no-unused-vars": "warn",
+      "prefer-const": "error",
+      "no-var": "error"
+    }
+  }
+];
+
+export default eslintConfig;
