+++ packages/config/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "@pilot/config",
+  "version": "0.1.0",
+  "private": true,
+  "type": "module",
+  "exports": {
+    "./eslint": "./eslint.config.mjs",
+    "./postcss": "./postcss.config.mjs",
+    "./tsconfig": "./tsconfig.base.json"
+  },
+  "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^8.56.0",
+    "@typescript-eslint/parser": "^8.56.0"
+  }
+}
